api/requirements.txt
supabase
google-genai
google-cloud-tasks
python-dotenv
google-cloud-bigquery
